+++ enterprise/admin/__init__.py
@@ -90,7 +90,7 @@ class EnterpriseCustomerEntitlementInline(admin.StackedInline):
 
     readonly_fields = ('ecommerce_coupon_url',)
     ecommerce_coupon_url.allow_tags = True
-    ecommerce_coupon_url.short_description = 'Coupon URL'
+    ecommerce_coupon_url.short_description = 'Seat Entitlement URL'
 
 
 class EnterpriseCustomerCatalogInline(admin.TabularInline):
@@ -112,14 +112,14 @@ class EnterpriseCustomerAdmin(DjangoObjectActions, SimpleHistoryAdmin):
     """
     list_display = (
         'name',
-        'uuid',
         'site',
         'active',
         'has_logo',
         'enable_dsc',
         'has_identity_provider',
         'has_enterprise_catalog',
-        'has_ecommerce_coupons'
+        'has_ecommerce_coupons',
+        'uuid',
     )
 
     list_filter = ('active',)